==> 1_Proteome Analysis.Rmd <==
---
title: "PISA_proteome_analysis"
author: "Christine Berryhill"
date: "2023-01-23"
output: html_document
---

# Purpose of this script is to filter out proteins with <HIGH FDR, fewer than 2 unique peptides, and contaminants. 
# This script also adds a pseudocount
# Reproducibility of the proteome


```{r Opening up the proteome files}
pisa_prot<- read.csv("Inputs/ProteinSummary.csv")

pisa_prot<- na.omit(pisa_prot)

library(dplyr)
library(ggplot2)
library(GGally)
library(stringr)
library(tidyverse)

```

# Filtering the proteome data and adding a pseudocount
```{r}
pisa_prot2<-pisa_prot%>%
  filter(X..Unique.Peptides >1)

pisa_prot2<- pisa_prot2%>%
  filter(Protein.FDR.Confidence..Combined =="High")

pisa_prot2<- pisa_prot2 %>%
  filter(!grepl("CONTAMINANT", Description))


pisa_work<- pisa_prot2[,c(4,24:39)]

rownames(pisa_work)<- pisa_work[,1]

pisa_work<- pisa_work[,-c(1)]

pisa_work<- pisa_work+0.1

pisa_work$Accession<- rownames(pisa_work)

pisa_work<- pisa_work[,c(17, 1:16)]
colnames(pisa_work)<- c("Accession", 
                        "Cont_Global_Rep1",
                        "Cont_Global_Rep2",
                        "Cont_Global_Rep3",
                        "Cont_Global_Rep4",
                         "Cont_PISA_Rep1",
                       "Cont_PISA_Rep2",
                       "Cont_PISA_Rep3",
                       "Cont_PISA_Rep4",
                       "DZNep_Global_Rep1",
                       "DZNep_Global_Rep2",
                       "DZNep_Global_Rep3",
                       "DZNep_Global_Rep4",
                        "DZNep_PISA_Rep1",
                       "DZNep_PISA_Rep2",
                       "DZNep_PISA_Rep3",
                       "DZNep_PISA_Rep4")



```

# Reproducibility analysis using GGpairs
```{r}
#Control Global Reps

control_global<- pisa_work[,c(2:5)]
a<-ggpairs(log10(control_global),
        title = "Correlation between Global Control Replicates",
        xlab = "log10(Abundance)",
        ylab = "log10(Abundance)",
        columnLabels = c("Rep 1", "Rep2","Rep 3", "Rep 4"),
        upper = list(continuous = wrap('cor', size = 5)),
        lower=list(continuous="smooth"))+
  theme_bw()
a
#DZNep Global Reps
dznep_global<- pisa_work[,c(10:13)]
b<-ggpairs(log10(dznep_global),
        title = "Correlation between Global DZNep Replicates",
        xlab = "log10(Abundance)",
        ylab = "log10(Abundance)",
        columnLabels = c("Rep 1", "Rep2","Rep 3", "Rep 4"),
        upper = list(continuous = wrap('cor', size = 5)),
        lower=list(continuous="smooth"))+
  theme_bw()
b
#Control PISA Reps
control_pisa<- pisa_work[,c(6:9)]
c<-ggpairs(log10(control_pisa),
        title = "Correlation between PISA Control Replicates",
        xlab = "log10(Abundance)",
        ylab = "log10(Abundance)",
        columnLabels = c("Rep 1", "Rep2","Rep 3", "Rep 4"),
        upper = list(continuous = wrap('cor', size = 5)),
        lower=list(continuous="smooth"))+
  theme_bw()
c
#DZNep PISA Reps

dznep_pisa<- pisa_work[,c(14:17)]
d<-ggpairs(log10(dznep_pisa),
        title = "Correlation between PISA DZNep Replicates",
        xlab = "log10(Abundance)",
        ylab = "log10(Abundance)",
        columnLabels = c("Rep 1", "Rep2","Rep 3", "Rep 4"),
        upper = list(continuous = wrap('cor', size = 5)),
        lower=list(continuous="smooth"))+
  theme_bw()
d

ggsave("Control_global_ggpair_correlation.eps", a, device = "eps", width = 7, height = 7)
ggsave("DZNep_global_ggpair_correlation.eps", b, device = "eps", width = 7, height = 7)
ggsave("Control_pisa_ggpair_correlation.eps", c, device = "eps", width = 7, height = 7)
ggsave("DZNep_pisa_ggpair_correlation.eps", d, device = "eps", width = 7, height = 7)

```


# Volcano plot of the proteome 
## Statistical analysis of the proteome
```{r Calculating FC and significance using a t test}


#determining the mean and sd of the normalized abundances
pisa_work$cont_global_mean<- rowMeans(pisa_work[,2:5])
pisa_work$cont_global_sd<-apply(pisa_work[,2:5], 1, sd)

pisa_work$dznep_global_mean<- rowMeans(pisa_work[,10:13])
pisa_work$dznep_global_sd<- apply(pisa_work[,10:13], 1, sd)


# Calculating log10FC

pisa_work$FC_global<- pisa_work$dznep_global_mean/pisa_work$cont_global_mean
#hist(pisa$FC) 
#histogram is not normal. So to do a t-test, need normal distribution. Therefore log transforming

pisa_work$log10FC_global<- log10(pisa_work$FC_global)
#hist(pisa$log10FC)
#now have a normal distribution

pisa_work<- na.omit(pisa_work)


#0 sites dropped

pisa_work<- pisa_work%>%
  rowwise() %>%
  mutate(pval = t.test(c(Cont_Global_Rep1, Cont_Global_Rep2, Cont_Global_Rep3, Cont_Global_Rep4),
                       c(DZNep_Global_Rep1, DZNep_Global_Rep2, DZNep_Global_Rep3, DZNep_Global_Rep4))$p.value)%>%
  ungroup()

#pisa_work$p.adjust<- p.adjust(pisa_work$pval, method = "BH")

pisa_work$log10.pvalue<- -log10(pisa_work$pval)

#desginating when something is up or down

pisa_work<- pisa_work%>%
  mutate(protein_type= case_when(FC_global > 1& log10.pvalue >=1.3 ~ "Up with DZNep",
                                 FC_global <= 1 & log10.pvalue >= 1.3 ~"Down with DZNep",
                                 TRUE ~ "ns"))



#obtain protein counts

pisa_work%>%
  count(protein_type)


```

```{r writing csv}
#write.csv(pisa_work, "Protome_global_changes.csv")

```
# Creating Volcano Plot
```{r volcano plot}

cols <- c("Up with DZNep" = "#ff1a1a", "Down with DZNep" = "#26b3ff", "ns" = "grey") 
sizes <- c("Up with DZNep" = 2, "Down with DZNep" = 2, "ns" = 1) 

a<-pisa_work%>%
  ggplot(aes(x = log2(FC_global),
             y = -log10(pval),
             fill = protein_type,    
             size = protein_type)) + 
  geom_point(shape = 21, # Specify shape and colour as fixed local parameters    
             colour = "black") + 
  geom_hline(yintercept = 1.3,
             linetype = "dashed") +
  scale_fill_manual(values = cols) + # Modify point colour
  scale_size_manual(values = sizes) + # Modify point size
  scale_x_continuous(      
                     limits = c(-2.5, 2.5)) +
  theme_classic()+
  labs(title= "Differential global protein abundances after DZNep treatment",
       x = "log2(DZNep/Control)",
       y = "-log10(pvalue)")+
  theme(legend.position = "none")

a
ggsave("Global proteome changes volcanoplot_3.24.eps", a, device = "eps", width = 5, height = 5)


```
