==> 10_ATP binding and all methyltransferase analysis.Rmd <==
```{r}
library(dplyr)
library(ggstatsplot)
library(ggplot2)

#Downloaded from Ensembl
atp<- read.csv("/home/cberry/Documents/PISA_Analysis/Analysis/QuickGo_ATPbinding.csv", sep = "\t")
methyl<- read.csv("/home/cberry/Documents/PISA_Analysis/Analysis/QuickGo_methyltransferases.csv", sep = "\t")

#Generated from Previously observed Kme site ananlysis.Rmd
pisa_hits<- read.csv("/home/cberry/Documents/PISA_Analysis/Analysis/Total workup of pisa data.csv")

#Generated from RNA seq analysis
rna_hits<- read.csv("/home/cberry/Documents/PISA_Analysis/Analysis/PISA and RNA data.csv")

```

```{r}
atp<- atp[!duplicated(atp$SYMBOL),]
atp_symbol<- as.vector(atp$SYMBOL)
methyl<- methyl[!duplicated(methyl$SYMBOL),]
methyl_symbol<- as.vector(methyl$SYMBOL)


pisa_hits<- pisa_hits%>%
  mutate(atp_binder = ifelse(Gene.Name %in% atp_symbol, "atp_binder", ""))

pisa_hits<- pisa_hits%>%
  mutate(methyltransferase = ifelse(Gene.Name %in% methyl_symbol, "methyltransferase", ""))

rna_hits<- rna_hits%>%
  mutate (atp_binder= ifelse(Gene.Name %in% atp_symbol, "atp_binder", ""))
rna_hits<- rna_hits%>%
  mutate (methyltransferase= ifelse(Gene.Name %in% methyl_symbol, "methyltransferase", ""))


print.chi<- function(x){ print(x)
  print("Expected") 
  print(x$expected) 
  print("Observed")
  print(x$observed) 
  print ("Residual")
  print(x$residual)}

chi.a<-chisq.test(table(pisa_hits$protein_type, pisa_hits$atp_binder))
chi.b<-chisq.test(table(pisa_hits$protein_type.pisa, pisa_hits$atp_binder))
chi.c<-chisq.test(table(pisa_hits$protein_type, pisa_hits$methyltransferase))
chi.d<-chisq.test(table(pisa_hits$protein_type.pisa, pisa_hits$methyltransferase))
chi.e<-chisq.test(table(rna_hits$gene_type, rna_hits$atp_binder))
chi.f<-chisq.test(table(rna_hits$gene_type, rna_hits$methyltransferase))



print.chi(chi.a)
print.chi(chi.b)
print.chi(chi.c)
print.chi(chi.d)
print.chi(chi.e)
print.chi(chi.f)
```

```{r}
a<-ggbarstats(
  data = pisa_hits,
  x = protein_type,
  y = atp_binder,
  package = "ggsci",
  palette = "default_jco"
) +
  labs(caption = NULL)
b<-ggbarstats(
  data = pisa_hits,
  x = protein_type.pisa,
  y = atp_binder,
  package = "ggsci",
  palette = "default_jco"
) +
  labs(caption = NULL)


c<-ggbarstats(
  data = pisa_hits,
  x = protein_type,
  y = methyltransferase,
  package = "ggsci",
  palette = "default_jco"
) +
  labs(caption = NULL)
d<-ggbarstats(
  data = pisa_hits,
  x = protein_type.pisa,
  y = methyltransferase,
  package = "ggsci",
  palette = "default_jco"
) +
  labs(caption = NULL)


ggsave("ATP Binder protein abundance.eps", a, device = "eps", width = 6, height = 6)
ggsave("ATP Binder pisa abundance.eps", b, device = "eps", width = 6, height = 6)
ggsave("methyltransferase protein abundance.eps",c, device = "eps", width = 6, height = 6)
ggsave("methyltransferase pisa abundance.eps", d, device = "eps", width = 6, height = 6)

```
